// front/src/core/apis/backend/generated/api.ts
export const AppArch = {
	Win32: "Win32",
	Win64: "Win64",
	LinuxDeb: "LinuxDeb",
	LinuxRpm: "LinuxRpm",
	LinuxSnap: "LinuxSnap",
} as const;

export type AppArch = typeof AppArch[keyof typeof AppArch];

/**
 *
 * @export
 * @interface AppMetadata
 */
export interface AppMetadata {
	/**
	 *
	 * @type {string}
	 * @memberof AppMetadata
	 */
	name: string;
	/**
	 *
	 * @type {AppVersion}
	 * @memberof AppMetadata
	 */
	version: AppVersion;
	/**
	 *
	 * @type {AppArch}
	 * @memberof AppMetadata
	 */
	arch: AppArch;
}

/**
 *
 * @export
 * @interface AppVersion
 */
export interface AppVersion {
	/**
	 *
	 * @type {number}
	 * @memberof AppVersion
	 */
	minor: number;
	/**
	 *
	 * @type {number}
	 * @memberof AppVersion
	 */
	major: number;
	/**
	 *
	 * @type {number}
	 * @memberof AppVersion
	 */
	revision: number;
	/**
	 *
	 * @type {string}
	 * @memberof AppVersion
	 */
	raw: string;
}